requirements.txt
# Core dependencies
numpy>=1.19.2
pillow>=8.0.1
tqdm>=4.62.0
matplotlib>=3.3.2
scikit-learn>=0.24.0
pandas>=1.3.0
ftfy
regex

# PyTorch - install separately based on your system
# For macOS: pip3 install torch torchvision
# For Linux/Windows with CUDA: pip3 install torch torchvision --index-url https://download.pytorch.org/whl/cu118

# CLIP - install after PyTorch
# pip install git+https://github.com/openai/CLIP.git